Added rechart instead of chartjs. Replaced table with bar chart.
